requirements.txt
streamlit
opencv-python
numpy
scikit-learn
scikit-image
matplotlib
Pillow
tensorflow
pydicom
plotly
streamlit-authenticator
pyyaml
python-multipart
pandas
seaborn
watchdog